+++ tests/cli_run_integration_tests.rs
@@ -49,18 +49,10 @@ mod cli_run_integration_tests {
                     overwrite: true,
                     append: false,
                 }),
-                #[cfg(feature = "parquet")]
-                parquet: None,
                 #[cfg(feature = "jsonl")]
                 jsonl: None,
                 #[cfg(feature = "sqlite")]
                 sqlite: None,
-                #[cfg(feature = "duckdb")]
-                duckdb: None,
-                #[cfg(feature = "postgres")]
-                postgres: None,
-                #[cfg(feature = "dm")]
-                dm: None,
             },
         };
 
@@ -109,18 +101,10 @@ mod cli_run_integration_tests {
                     overwrite: true,
                     append: false,
                 }),
-                #[cfg(feature = "parquet")]
-                parquet: None,
                 #[cfg(feature = "jsonl")]
                 jsonl: None,
                 #[cfg(feature = "sqlite")]
                 sqlite: None,
-                #[cfg(feature = "duckdb")]
-                duckdb: None,
-                #[cfg(feature = "postgres")]
-                postgres: None,
-                #[cfg(feature = "dm")]
-                dm: None,
             },
         };
 
@@ -210,18 +194,10 @@ mod cli_run_integration_tests {
                     overwrite: true,
                     append: false,
                 }),
-                #[cfg(feature = "parquet")]
-                parquet: None,
                 #[cfg(feature = "jsonl")]
                 jsonl: None,
                 #[cfg(feature = "sqlite")]
                 sqlite: None,
-                #[cfg(feature = "duckdb")]
-                duckdb: None,
-                #[cfg(feature = "postgres")]
-                postgres: None,
-                #[cfg(feature = "dm")]
-                dm: None,
             },
         };
 
@@ -257,18 +233,10 @@ mod cli_run_integration_tests {
             exporter: ExporterConfig {
                 #[cfg(feature = "csv")]
                 csv: None,
-                #[cfg(feature = "parquet")]
-                parquet: None,
                 #[cfg(feature = "jsonl")]
                 jsonl: None,
                 #[cfg(feature = "sqlite")]
                 sqlite: None,
-                #[cfg(feature = "duckdb")]
-                duckdb: None,
-                #[cfg(feature = "postgres")]
-                postgres: None,
-                #[cfg(feature = "dm")]
-                dm: None,
             },
         };
 
@@ -302,18 +270,10 @@ mod cli_run_integration_tests {
                     overwrite: true,
                     append: false,
                 }),
-                #[cfg(feature = "parquet")]
-                parquet: None,
                 #[cfg(feature = "jsonl")]
                 jsonl: None,
                 #[cfg(feature = "sqlite")]
                 sqlite: None,
-                #[cfg(feature = "duckdb")]
-                duckdb: None,
-                #[cfg(feature = "postgres")]
-                postgres: None,
-                #[cfg(feature = "dm")]
-                dm: None,
             },
         };
 
@@ -348,18 +308,10 @@ mod cli_run_integration_tests {
                     overwrite: true,
                     append: false,
                 }),
-                #[cfg(feature = "parquet")]
-                parquet: None,
                 #[cfg(feature = "jsonl")]
                 jsonl: None,
                 #[cfg(feature = "sqlite")]
                 sqlite: None,
-                #[cfg(feature = "duckdb")]
-                duckdb: None,
-                #[cfg(feature = "postgres")]
-                postgres: None,
-                #[cfg(feature = "dm")]
-                dm: None,
             },
         };
 
@@ -435,18 +387,10 @@ mod cli_run_integration_tests {
                     overwrite: true,
                     append: false,
                 }),
-                #[cfg(feature = "parquet")]
-                parquet: None,
                 #[cfg(feature = "jsonl")]
                 jsonl: None,
                 #[cfg(feature = "sqlite")]
                 sqlite: None,
-                #[cfg(feature = "duckdb")]
-                duckdb: None,
-                #[cfg(feature = "postgres")]
-                postgres: None,
-                #[cfg(feature = "dm")]
-                dm: None,
             },
         };